+++ cypress/integration/unsubscribe.spec.js
@@ -37,30 +37,58 @@ describe("unsubscribe page", () => {
     cy.get('[data-cy="error-box-items"').children().should("have.length", 1);
   });
 
-  it("Redirects to thank you page on successful submit", () => {
-    cy.intercept("POST", "/api/**", {
-      statusCode: 201,
-      body: {
-        data: "test success",
-      },
-    });
+  it("Test unsubscribe on successful submit", () => {
+    cy.intercept("/api/**", { fixture: "signup" }).as("response");
 
     cy.get('[id="email"]').type("some@email.com");
     cy.get('[data-cy="unsubscribe-submit"]').click();
-    cy.get('[data-cy="toggle-language-link"]').then(($el) => {
-      const text = $el.text();
-      if (text === "English") {
-        cy.url().should(
-          "contains",
-          "/merci?e=s***%40****l.***&ref=unsubscribe"
-        );
-      }
-      if (text === "Français") {
-        cy.url().should(
-          "contains",
-          "/thankyou?e=s***%40****l.***&ref=unsubscribe"
-        );
-      }
+    cy.url().should("contains", "/thankyou?e=s***%40****l.***&ref=unsubscribe");
+
+    cy.wait("@response").then((xhr) => {
+      cy.log(xhr);
+      expect(xhr.request.method).to.equal("POST");
+      expect(xhr.response.statusCode).to.equal(200);
+      expect(xhr.response).to.have.property("headers");
+      expect(xhr.response).to.have.property("body");
+
+      // redirect to thankyou page on successful submit
+      cy.url().should(
+        "contains",
+        "/thankyou?e=s***%40****l.***&ref=unsubscribe"
+      );
+
+      // confirm email
+      expect(xhr.response.body.email).to.equal("some@email.com");
+    });
+  });
+
+  it("Test 400 bad resquest", () => {
+    cy.intercept("/api/**", { statusCode: 400 }).as("response");
+
+    cy.get('[id="email"]').type("some@email.com");
+    cy.get('[data-cy="unsubscribe-submit"]').click();
+
+    cy.wait("@response").then((xhr) => {
+      cy.log(xhr);
+      expect(xhr.request.method).to.equal("POST");
+      expect(xhr.response.statusCode).to.equal(400);
+      cy.get('[data-cy="error-box"').should("exist");
+      cy.get('[data-cy="error-box"').should("contain.text", "can’t be found");
+    });
+  });
+
+  it("Server return 500 bad response", () => {
+    cy.intercept("/api/**", { statusCode: 500 }).as("response");
+
+    cy.get('[id="email"]').type("some@email.com");
+    cy.get('[data-cy="unsubscribe-submit"]').click();
+
+    cy.wait("@response").then((xhr) => {
+      cy.log(xhr);
+      expect(xhr.request.method).to.equal("POST");
+      expect(xhr.response.statusCode).to.equal(500);
+      cy.get('[data-cy="error-box"').should("exist");
+      cy.get('[data-cy="error-box"').should("contain.text", "unknown error");
     });
   });
 });