+++ CHANGELOG.md
@@ -9,6 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Changed
 
+- test-signup-unsubscribe-flow
+- Updated e2e test for signup and unsubscribe workflow
 - Change color on confirmation when feedback submited successfully for feedback component
 - Updated content on the experiments, about and sign up pages
 - Updated French path in French language